+++ Astral Clash/Assets/MusicVolume.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 2a4b54118d7b48d42b1b0f4faf665306
-timeCreated: 1446318705
+guid: 31743b0401679414cbe84dd934cc0b18
+timeCreated: 1446316561
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
