+++ Astral Clash/Assets/Scripts/CharSelect.cs
@@ -1,20 +1,15 @@
 using UnityEngine;
 using System.Collections;
+using UnityEngine.UI;
 
-public class CharSelect : MonoBehaviour {
+public class CharSelect : Menu {
 
 	public Sprite[,] CharSprite = new Sprite[2, 2];
 	public Sprite[] sprites;
-	public int selected;
 	public int selected2;
-	public int max;
 	public int max2;
-	public bool axisPressed;
-	public bool axisPressedLR;
-	public Match match;
 	private int Players;
-	public TextMesh pText;
-	public GameObject prevMenu;
+	public Text pText;
 
 	// Use this for initialization
 	void Start () {
@@ -26,88 +21,6 @@ public class CharSelect : MonoBehaviour {
 		selected = 0;
 		selected2 = 0;
 		Players = 1;
-	
-	}
-	
-	// Update is called once per frame
-	void Update () {
-
-		if (axisPressed == false) {
-			if (Input.GetAxis ("MenuDPad") < 0) {
-				
-				if (selected == max) {
-					
-					selected = 0;
-					
-				} else {
-					selected++;
-				}
-				
-				//selectionEffect ();
-				axisPressed = true;
-				
-			} else if (Input.GetAxis ("MenuDPad") > 0) {
-				
-				if (selected == 0) {
-					
-					selected = max;
-					
-				} else {
-					selected--;
-				}
-				
-				//selectionEffect ();
-				axisPressed = true;
-				
-			}
-		}
-		
-		if (axisPressedLR == false) {
-			if (Input.GetAxis ("MenuLR") > 0) {
-
-				if (selected2 == max2) {
-					
-					selected2 = 0;
-					
-				} else {
-					selected2++;
-				}
-
-				//selectionEffect ();
-				axisPressedLR = true;
-				
-			} else if (Input.GetAxis ("MenuLR") < 0) {
-
-				if (selected2 == 0) {
-					
-					selected2 = max2;
-					
-				} else {
-					selected2--;
-				}
-				//selectionEffect ();
-				axisPressedLR = true;
-				
-			}
-		}
-		
-		if (Input.GetAxis ("MenuDPad") == 0) {
-			
-			axisPressed = false;
-			
-		}
-		
-		if (Input.GetAxis ("MenuLR") == 0) {
-			
-			axisPressedLR = false;
-			
-		}
-
-		if(Input.GetButtonDown("Cancel")){
-			
-			BackMenu();
-			
-		}
 	
 	}
 
@@ -174,21 +87,6 @@ public class CharSelect : MonoBehaviour {
 			GameObject.Find("GameManager").GetComponent<GameManager>().CreateNewMatch(match);
 		
 		}
-		
-	}
-
-	public void SetMatch(Match newMatch){
-		match = newMatch;
-		
-	}
-
-	void BackMenu(){
-		
-		prevMenu.SetActive (true);
-		Camera.current.transform.position = new Vector3 (160, 0, -10);
-		Players = 1;
-		pText.text = "Player "+Players.ToString();
-		this.gameObject.SetActive (false);
 		
 	}
